CODE: Replace floating point arithmetic with integer arithmetic in many parts of the code to attempt to eliminate problems with networking between differently compiled binaries. CHANGE: There is no longer a generic tendency against using private cars for longer journies. FIX: Trains and other railed vehicles would sometimes get stuck when signals were deleted.

+++ besch/fabrik_besch.h
@@ -226,7 +226,7 @@ private:
 	uint16 produkte; //"products" (Babelfish)
 	uint8 fields;	// only if there are any ...
 	uint16 pax_level;
-	float electricity_proportion; // Modifier of electricity consumption.
+	uint16 electricity_proportion; // Modifier of electricity consumption.
 	uint16 inverse_electricity_proportion;
 	bool electricity_producer;
 	uint8 upgrades; // The industry types to which this industry can be upgraded.
@@ -272,7 +272,7 @@ public:
 	/* level for post and passenger generation */
 	int get_pax_level() const { return pax_level; }
 
-	float get_electricity_proportion() const { return electricity_proportion; }
+	uint16 get_electricity_proportion() const { return electricity_proportion; }
 	uint16 get_inverse_electricity_proportion() const { return inverse_electricity_proportion; }
 
 	int is_electricity_producer() const { return electricity_producer; }
